Check-In - Dynamically finding port name
Port number keeps on changing whenever we plugin hardware to laptop.
This checkin will try to find active ports name match with prolific and
use that for rest of the earlier code.

Currently the string prolific is hardcoded. We can make it dynamic by
letting user enter that in console.